refactor: return connection to DB by ts config
