+++ .env
@@ -15,17 +15,17 @@ PGDATA=/var/lib/pgsql/data
 #
 TIME_ZONE=Europe/Minsk
 
-# TYPEORM_CONFIG
-TYPEORM_CONNECTION=postgres
-TYPEORM_HOST=host.docker.internal
-TYPEORM_USERNAME=postgres
-TYPEORM_PASSWORD=123321
-TYPEORM_DATABASE=postgres
-TYPEORM_PORT=5432
-TYPEORM_SYNCHRONIZE=false
-TYPEORM_LOGGING=true
-TYPEORM_ENTITIES=src/entities/*.ts
-TYPEORM_ENTITIES_DIR=src/entities
-TYPEORM_MIGRATIONS_DIR=src/migration
-TYPEORM_MIGRATIONS_RUN=true
-TYPEORM_MIGRATIONS=src/migration/*.ts
+## TYPEORM_CONFIG
+#TYPEORM_CONNECTION=postgres
+#TYPEORM_HOST=host.docker.internal
+#TYPEORM_USERNAME=postgres
+#TYPEORM_PASSWORD=123321
+#TYPEORM_DATABASE=postgres
+#TYPEORM_PORT=5432
+#TYPEORM_SYNCHRONIZE=false
+#TYPEORM_LOGGING=true
+#TYPEORM_ENTITIES=src/entities/*.ts
+#TYPEORM_ENTITIES_DIR=src/entities
+#TYPEORM_MIGRATIONS_DIR=src/migration
+#TYPEORM_MIGRATIONS_RUN=true
+#TYPEORM_MIGRATIONS=src/migration/*.ts
