+++ package.json
@@ -16,7 +16,7 @@
     "doc": "jsdoc -r ./src",
     "docker:up": "docker-compose up --build",
     "docker:down": "docker-compose down",
-    "typeorm": "ts-node --files -r tsconfig-paths/register ./node_modules/typeorm/cli.js --config src/common/111ormconfig.ts",
+    "typeorm": "ts-node --files -r tsconfig-paths/register ./node_modules/typeorm/cli.js --config src/common/ormconfig.ts",
     "db:create": "npm run typeorm migration:generate -n TodoMigrations",
     "db:run": "npm run typeorm -- migration:run",
     "migrationg": "ts-node ./node_modules/typeorm/cli.js migration:generate -n N"
