+++ src/server.ts
@@ -1,8 +1,9 @@
+import { createConnection } from 'typeorm';
 import { env } from './common/config';
 import app from './app';
-import { connect } from './memoryDb/postgresDB';
+import config from './common/ormconfig';
 
-connect()
+createConnection(config)
   .then(() => {
     app.listen(env.PORT, () =>
       process.stdout.write(`App is running on http://localhost:${env.PORT}\n`)
